:art: Style: 프로젝트 섹션 컴포넌트 분리
Style: 프로젝트 섹션 컴포넌트 분리
