+++ app/project/[projectId]/_components/Section/TeamMemberSection.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function TeamMemberSection() {
+  return (
+    <section className="mt-10">
+      <p className="mb-4 text-lg font-bold">팀원</p>
+    </section>
+  );
+}
+
+export default TeamMemberSection;
